requirements.txt
fastapi
uvicorn
requests
sseclient-py
python-dotenv
